fixed incompatible types

+++ src/soundm4a.h
@@ -11,7 +11,7 @@ extern const ma_data_source_vtable m4a_file_data_source_vtable;
 
 void m4a_on_audio_frames(ma_device *pDevice, void *pFramesOut, const void *pFramesIn, ma_uint32 frameCount);
 
-MA_API ma_result m4a_read_pcm_frames_wrapper(void *pDecoder, void* pFramesOut, long unsigned int frameCount, long unsigned int* pFramesRead);
+MA_API ma_result m4a_read_pcm_frames_wrapper(void *pDecoder, void* pFramesOut, size_t frameCount, size_t *pFramesRead);
 
 MA_API ma_result m4a_seek_to_pcm_frame_wrapper(void *pDecoder, long long int frameIndex, ma_seek_origin origin);
